pyproject.toml
[tool.poetry]
name = "test-task-student-score-collector"
version = "0.1.0"
description = ""
authors = ["Julia Bengert <yubengert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.0"
uvicorn = {extras = ["standard"], version = "^0.37.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
pydantic-settings = "^2.11.0"
asyncpg = "^0.30.0"
alembic = "^1.16.5"
pyjwt = "^2.10.1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.20"
aiogram = "^3.22.0"
vkbottle = "^4.6.2"


[tool.poetry.group.develop.dependencies]
black = "^25.9.0"
isort = "^6.1.0"
flake8 = "^7.3.0"


[tool.black]
line-length = 119

[tool.isort]
line_length = 119


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
